18.12.22

+++ app.py
@@ -69,7 +69,7 @@ class Order(db.Model):
 
 class Offer(db.Model):
     __tablename__ = "offer"
-    id = db.Column(db.Integer(100), primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     order_id = db.Column(db.Integer, db.ForeignKey(f"{Order.__tablename__}.id"))
     executor_id = db.Column(db.Integer, db.ForeignKey(f"{User.__tablename__}.id"))
